admin/live: show running-query duration in the Live view
Adds an elapsed-time column to the running-queries table so operators can
see how long each in-flight statement has been running, not just its
progress.

- QueryStatus.ElapsedMS: how long the current statement has run, computed
  on the owning CP from the connection's query-start (server.ConnDetailByPID).
  It's per-item, so it survives the cross-CP /queries merge unchanged; 0 when
  the session is idle.
- RunningQueries populates it from the live conn.
- UI: a "Duration" column in the Live table; shared fmtDuration() in
  lib/format.ts (replaces the dialog's local fmtElapsed, now reused by both
  the column and the detail dialog).
- harness admin_query_detail: asserts the in-flight query's /queries list
  item reports a positive elapsed_ms (the field is live end-to-end).
